separated encruption functions in encryption.ts file and sendFile function in worker file
