+++ client/app/page.tsx
@@ -7,6 +7,9 @@ import './style.css'
 import JSEncrypt from 'jsencrypt';
 import JSON from 'json3';
 import pako from 'pako';
+import * as crpt from './encryption';
+
+
 
 const { subtle } = globalThis.crypto;
 
@@ -49,7 +52,7 @@ export default function Home() {
   
   useEffect(() => {
     if (isConnecting)
-      generateKeyPair();
+    generateKeyPair();
   }, [isConnecting]);
 
   useEffect(() => {
@@ -74,6 +77,11 @@ export default function Home() {
     setPeerAccepted(false);
   }
 
+
+  const generateKeyPair = async () => {
+    setKeyPair(await crpt.generateKeyPair());
+  }
+
   const generatePeerName = () => {
     let adjectives = ['Amazing', 'Awesome', 'Beautiful', 'Brave', 'Bright', 'Calm', 'Clever', 'Cool', 'Cute', 'Dazzling', 'Elegant', 'Enchanting', 'Fabulous', 'Fantastic', 'Friendly', 'Funny', 'Gentle', 'Glamorous', 'Gorgeous', 'Graceful', 'Handsome', 'Happy', 'Healthy', 'Helpful', 'Hilarious', 'Humorous', 'Jolly', 'Joyous', 'Kind', 'Lively', 'Lovely', 'Lucky', 'Magnificent', 'Nice', 'Perfect', 'Pleasant', 'Proud', 'Silly', 'Smiling', 'Splendid', 'Successful', 'Thoughtful', 'Victorious', 'Vivacious', 'Witty', 'Wonderful','Adventurous', 'Blissful', 'Charming', 'Delightful', 'Exquisite', 'Fierce', 'Glowing', 'Harmonious', 'Incredible', 'Jubilant'];
 
@@ -90,177 +98,6 @@ export default function Home() {
     console.log(e.target.files[0].size);
     console.log(e.target.files[0].type);
   };
-
-  const sendFile = async (file: any) => {
-    const CHUNK_SIZE = 5 * 1024 * 1024; // 5 MB
-    const reader = new FileReader();
-  
-    if (file) {
-      reader.onload = async () => {
-        const base64Data = reader.result! as string;
-        console.log("Original file: ", base64Data.length);
-        let start = new Date;
-        const compressedData = pako.deflate(base64Data);
-        let end = new Date;
-        console.log("Compression time: ", (end.getTime() - start.getTime())/1000);
-        console.log("Compressed file: ", compressedData.length);
-  
-        const totalChunks = Math.ceil(compressedData.length / CHUNK_SIZE);
-        let totalSize = 0;
-        for (let i = 0; i < totalChunks; i++) {
-          let sendStart = new Date;
-          const start = i * CHUNK_SIZE;
-          const end = Math.min(start + CHUNK_SIZE, compressedData.length);
-          const chunk = compressedData.slice(start, end);
-          totalSize += chunk.length;
-          console.log("Chunk size ", i, ": ", chunk.length);
-  
-          const key = await generateAESKey();
-          let startEncr = new Date;
-          const encryptedChunk = await encryptAES(chunk.buffer, key);
-          let endEncr = new Date;
-          console.log("Encryption time", i, ": ", (endEncr.getTime() - startEncr.getTime())/1000);
-          const exportedKey = await crypto.subtle.exportKey("raw", key);
-          const stringKey = arrayBufferToString(exportedKey);
-  
-          const encryptedData = JSON.stringify({
-            name: file.name,
-            key: await encrypt(stringKey),
-            iv: arrayBufferToString(encryptedChunk.iv.buffer),
-            chunk: arrayBufferToString(encryptedChunk.encryptedData),
-            totalChunks: totalChunks,
-            currentChunk: i + 1,
-          });
-  
-          socket!.send("file: " + encryptedData);
-          let sendEnd = new Date;
-          console.log("Send time", i, ": ", (sendEnd.getTime() - sendStart.getTime())/1000);
-        }
-        console.log("Total size: ", totalSize);
-      };
-  
-      reader.readAsDataURL(file);
-      reader.onerror = (error) => {
-        console.log("error: ", error);
-      };
-    }
-  };
-  
-  function arrayBufferToString(buffer: ArrayBuffer): string {
-    const view = new Uint8Array(buffer);
-    const length = view.length;
-    const result = new Array(length);
-  
-    for (let i = 0; i < length; i++) {
-      result[i] = String.fromCharCode(view[i]);
-    }
-  
-    return result.join('');
-  }  
-
-function stringToArrayBuffer(str: string): ArrayBuffer {
-  const view = new Uint8Array(str.length);
-
-  for (let i = 0; i < str.length; i++) {
-    view[i] = str.charCodeAt(i);
-  }
-  
-  return view.buffer;
-}
-  
-
-// Generate a random AES key
-async function generateAESKey() {
-  return await crypto.subtle.generateKey(
-    {
-      name: "AES-CBC",
-      length: 256, // Key size in bits
-    },
-    true, // Extractable key
-    ["encrypt", "decrypt"]
-  );
-}
-
-// Encrypt data using AES
-async function encryptAES(data: ArrayBuffer, key: CryptoKey): Promise<{ encryptedData: ArrayBuffer, iv: Uint8Array }>  {
-  // Generate an initialization vector (IV)
-  const iv = crypto.getRandomValues(new Uint8Array(16));
-
-  // Encrypt the data with AES-CBC
-  const encryptedData = await crypto.subtle.encrypt(
-    {
-      name: "AES-CBC",
-      iv: iv,
-    },
-    key,
-    data
-  );
-
-  // Return the encrypted data and IV as an object
-  return {
-    encryptedData: encryptedData,
-    iv: iv,
-  };
-}
-
-async function decryptAES(encryptedData: BufferSource, iv : BufferSource , key: CryptoKey): Promise<ArrayBuffer> {
-  // Decrypt the data with AES-CBC
-  const decryptedData = await crypto.subtle.decrypt(
-    {
-      name: "AES-CBC",
-      iv: iv,
-    },
-    key,
-    encryptedData
-  );
-
-  return decryptedData;
-}
-  
-  const encrypt = async (message: string) => {
-    let publicKeyBuffer = await subtle.exportKey("spki", peerPublicKey!)
-    let publicKey = Buffer.from(publicKeyBuffer).toString('base64');
-    const jsEncrypt = new JSEncrypt();
-    jsEncrypt.setPublicKey(publicKey);
-    return jsEncrypt.encrypt(message);
-  }
-
-  const decrypt = async (message: string) => {
-    let privateKeyBuffer = await subtle.exportKey("pkcs8", keyPair!.privateKey)
-    let privateKey = Buffer.from(privateKeyBuffer).toString('base64');
-    const jsEncrypt = new JSEncrypt();
-    jsEncrypt.setPrivateKey(privateKey);
-    return jsEncrypt.decrypt(message);
-  }
-
-  const generateKeyPair = async () => {
-    let localKeyPair = await subtle.generateKey(
-      {
-        name: "RSA-OAEP",
-        modulusLength: 2048,
-        publicExponent: new Uint8Array([1, 0, 1]),
-        hash: "SHA-256",
-      },
-      true,
-      ["encrypt", "decrypt"]
-    );
-    setKeyPair(localKeyPair);
-    console.log('Generated key pair: ', localKeyPair);
-  }
-
-
-  const sendPublicKey = async (socket: WebSocket) => {
-    let publicKeyBuffer = await subtle.exportKey("spki", keyPair!.publicKey);
-    let publicKey = Buffer.from(publicKeyBuffer).toString('base64');
-    socket.send('public-key: ' + publicKey);
-  }
-
-
-  const importAndSetPeerPublicKey = async (key: string) => {
-    let publicKey = await subtle.importKey("spki", Buffer.from(key, 'base64'), {name: "RSA-OAEP", hash: "SHA-256"}, true, ["encrypt"]);
-    setPeerPublicKey(publicKey);
-    console.log('Imported public key: ', publicKey);
-  }
 
 
   const getUrl = () => {
@@ -287,12 +124,12 @@ async function decryptAES(encryptedData: BufferSource, iv : BufferSource , key: 
 
       } else if (event.data.slice(0, 13) === 'client-info: ') {
         let encrypted = event.data.slice(13);
-        let decrypted = await decrypt(encrypted);
+        let decrypted = await crpt.decrypt(encrypted, keyPair);
         console.log('Decrypted client info: ', decrypted);
         setPairInfo(decrypted);
 
       } else if (event.data === 'connected') {
-        sendPublicKey(socket!);
+        crpt.sendPublicKey(socket!, keyPair);
 
       } else if (event.data.slice(0,11) === 'client-id: '){
         let clientId = event.data.slice(11);
@@ -302,14 +139,14 @@ async function decryptAES(encryptedData: BufferSource, iv : BufferSource , key: 
       } else if (event.data.slice(0, 12) === 'public-key: ') {
         console.log('Received public key: ', event.data.slice(12));
         let key = event.data.slice(12);
-        importAndSetPeerPublicKey(key);
+        crpt.importAndSetPeerPublicKey(key);
 
       } else if (event.data.slice(0, 5) === 'msg: ') {
         // Q: Why use msg: prefix?
         // A: Maybe someone sends a message that starts with 'client-id: ' or 'public-key: '?
         let encrypted = event.data.slice(5);
 
-        let decryptedString = await decrypt(encrypted);
+        let decryptedString = await crpt.decrypt(encrypted, keyPair);
         setHistory((prev: any) => [...prev, 'Peer: ' + decryptedString]);
         setChatStarted(true);
       } else if(event.data === 'accept-pairing'){
@@ -343,7 +180,7 @@ async function decryptAES(encryptedData: BufferSource, iv : BufferSource , key: 
       var peerName = generatePeerName();
       //setPeerName(generatePeerName());
     //   let encryptedString = await encrypt("device: " + navigator.userAgent);
-      let encryptedString = await encrypt(peerName);
+      let encryptedString = await crpt.encrypt(peerName, peerPublicKey);
       setClientName(peerName);
       socket!.send("client-info: " + encryptedString);
       console.log('Sent client info: ', encryptedString);
@@ -356,11 +193,11 @@ async function decryptAES(encryptedData: BufferSource, iv : BufferSource , key: 
         name: 'AES-CBC',
         length: 256
       };
-      const decryptedKey = await decrypt(json.key) as string;
+      const decryptedKey = await crpt.decrypt(json.key, keyPair) as string;
       console.log("1");
-      const key = await crypto.subtle.importKey('raw', stringToArrayBuffer(decryptedKey), algorithm, true, ['encrypt', 'decrypt']);
+      const key = await crypto.subtle.importKey('raw', crpt.stringToArrayBuffer(decryptedKey), algorithm, true, ['encrypt', 'decrypt']);
       console.log("2");
-      const decryptedChunk = await decryptAES(stringToArrayBuffer(json.chunk), stringToArrayBuffer(json.iv), key);
+      const decryptedChunk = await crpt.decryptAES(crpt.stringToArrayBuffer(json.chunk), crpt.stringToArrayBuffer(json.iv), key);
       console.log("3");
       // Append the received chunk to the file
       let combinedLength = receivedChunks.byteLength + decryptedChunk.byteLength;
@@ -411,7 +248,7 @@ async function decryptAES(encryptedData: BufferSource, iv : BufferSource , key: 
     
     setHistory((prev: any) => [...prev, 'You: ' + text]);
 
-    let encryptedString = await encrypt(text);
+    let encryptedString = await crpt.encrypt(text, peerPublicKey);
     socket.send("msg: " + encryptedString);
     setMessage('');
     setChatStarted(true);
@@ -423,8 +260,23 @@ async function decryptAES(encryptedData: BufferSource, iv : BufferSource , key: 
   };
 
   const handleSendFile = async () => {
-    await sendFile(selectetdFile[0]);
+    
+  const worker = new Worker('/send-file.ts');
+
+  // Define the event listener to handle messages from the worker
+  worker.onmessage = function (event) {
+    socket!.send(event.data);
   };
+
+  // Send data to the worker
+  const data = {file: selectetdFile[0], peerPublicKey: peerPublicKey};
+  worker.postMessage(data);
+
+  // Cleanup the worker when the component unmounts
+  return () => {
+    worker.terminate();
+  };
+};
 
 
   const clearMessages = () => {
